requirements.txt
MarkupSafe==0.15
argparse==1.2.1
distribute==0.6.24
python-dateutil==2.1
pytz==2012j
six==1.2.0
wsgiref==0.1.2
